[Variant] Move `ArrayVariantToArrowRowBuilder` to `variant_to_arrow` (#9094)
# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.
-->

- Part of #8082.

# Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

To support `Lists/Array` in `variant_get`, it's better to move
`ArrayVariantToArrowRowBuilder` from `shred_variant` to
`variant_to_arrow` and be shared with `variant_get`. In the meantime,
some code movement in `variant_to_arrow` would help to get a better
overview of the overall implementation

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

This PR can be reviewed commit by commit:
- Move `VariantToArrowRowBuilder` and related impl to the top of
`variant_to_arrow`
- Push the `FixedSizeList` check from
`make_variant_to_shredded_variant_arrow_row_builder` down to
`ArrayVariantToArrowRowBuilder`
- Move `ArrayVariantToArrowRowBuilder` to `variant_to_arrow`

# Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

Covered by existing tests

# Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.
-->

No
